requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-storage==1.43.0
pandas==1.3.5
fsspec==2022.1.0